backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
yt-dlp>=2024.10.22
chromadb>=0.5.0
anthropic>=0.40.0
openai>=1.40.0
google-generativeai>=0.8.0
python-dotenv==1.0.1
aiosqlite==0.19.0
pydantic>=2.12.0
youtube-transcript-api==0.6.2
